# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "High-performance video streaming and processing API built with Go",
        "title": "Video Streaming API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8001",
    "paths": {
...
        "/api/v2/videos/{platform}/{video_id}": {
            "get": {
                "description": "Retrieve video metadata from various platforms",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Get video information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Platform (youtube, bilibili, twitter, instagram, twitch)",
                        "name": "platform",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Video ID",
                        "name": "video_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.VideoInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "models.Format": {
            "type": "object",
            "properties": {
                "bitrate": {
                    "type": "integer"
                },
                "codec": {
                    "type": "string"
                },
                "ext": {
                    "type": "string"
                },
                "filesize": {
                    "type": "integer"
                },
                "format_id": {
                    "type": "string"
                },
                "quality": {
                    "type": "string"
                },
                "resolution": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "models.VideoInfo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "formats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Format"
                    }
                },
                "id": {
                    "type": "string"
                },
                "like_count": {
                    "type": "integer"
                },
                "platform": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "upload_date": {
                    "type": "string"
                },
                "uploader": {
                    "type": "string"
                },
                "view_count": {
                    "type": "integer"
                }
            }
        }
    }
}